refactor(crypto): Move store::{CryptoStoreError, Result} into new module
